Even more cargo tyupes

+++ src/script/api/script_types.hpp
@@ -111,7 +111,7 @@
 
 /* Define all types here, so they are added to the API docs. */
 typedef uint BridgeID;         ///< The ID of a bridge type.
-typedef uint8_t CargoID;       ///< The ID of a cargo.
+typedef uint16_t CargoID;       ///< The ID of a cargo.
 typedef uint16_t EngineID;     ///< The ID of an engine.
 typedef uint16_t GoalID;       ///< The ID of a goal.
 typedef uint16_t GroupID;      ///< The ID of a group.
